Add view problem component and refactor functionality from previous create problem component
