+++ App.js
@@ -1,12 +1,12 @@
 import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Text, View } from 'react-native';
-import CreateProblem from './app/Pages/CreateProblem';
+import ViewProblems from './app/Pages/ViewProblems';
 
 export default function App() {
   return (
     <View style={styles.container}>
       <Text>Open up App.js to start working on your app!</Text>
-      <CreateProblem />
+      <ViewProblems />
       <StatusBar style="auto" />
     </View>
   );
